sample: link app and library project
