+++ sample/build.gradle
@@ -43,6 +43,7 @@ android {
 }
 
 dependencies {
+    implementation project(path: ':muffin')
     implementation 'androidx.appcompat:appcompat:1.4.2'
     implementation 'com.google.android.material:material:1.6.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
