+++ sample/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
     package="dev.luncliff.sample">
 
     <application
-        android:name=".App"
+        android:name=".MuffinApp"
         android:allowBackup="false"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
